Rearrange order when halfway through other pane
Previously, the panes would be reordered only when the top of the
dragged pane crossed the top of another pane.  

With this change, a pane will be reordered when its top crosses halfway
over the previous pane when being dragged up/left, or when the bottom
passes halfway over the pane below/right.