Fix nasty bug where all Panes could end up sharing the same static style
If you don't specify an explicit style prop for Panes, then the Object.assign
in the motion cycle assigns to the static default {} object shared by all
Panes.  This causes all the Panes to jump to the same position and overlap.
This seems only noticeable if you disable effects.  The fix is to clone
the Pane's style property rather than direct-manipulating it.

Also remove a needless spring interpolation if you have disabled effects.